school db created and populated
