+++ src/grdb/database/models/__init__.py
@@ -12,3 +12,4 @@ from grdb.database.models.raman_file import RamanFile
 from grdb.database.models.raman_analysis import RamanAnalysis
 from grdb.database.models.properties import Properties
 from grdb.database.models.user import User
+from grdb.database.models.institution import Institution
