+++ src/grdb/server/app/views/__init__.py
@@ -1,3 +1,3 @@
 # blueprints
 from .auth import auth
-from .index import index
+from .experiments import experiments
