+++ src/requirements.txt
@@ -5,3 +5,4 @@ SQLAlchemy==1.3.23
 Werkzeug==2.0.2
 PyMySQL==1.0.2
 PyJWT==1.4.2
+gunicorn==20.1.0
